code refactor

+++ web-crawler-core/src/main/java/org/rc/webcrawler/lib/WebPageHandler.java
@@ -14,16 +14,16 @@ class WebPageHandler {
 
     private static final Logger logger = Logger.getLogger(WebPageHandler.class.getName());
 
-    public static final BiFunction<String, Long, Optional<Connection.Response>> PAGE_FETCHER = WebPageHandler::fetch;
+    public static final BiFunction<String, Integer, Optional<Connection.Response>> PAGE_FETCHER = WebPageHandler::fetch;
     public static final BiFunction<Connection.Response, Predicate<String>, Optional<Stream<String>>> URL_EXTRACTOR = WebPageHandler::extract;
 
 
-    private static Optional<Connection.Response> fetch(String url, long timeoutInMills) {
+    private static Optional<Connection.Response> fetch(String url, int timeoutInMills) {
         try {
             return Optional.of(
                     Jsoup.connect(url)
                             .userAgent("Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/535.21 (KHTML, like Gecko) Chrome/19.0.1042.0 Safari/535.21")
-                            .timeout((int) timeoutInMills)
+                            .timeout(timeoutInMills)
                             .ignoreContentType(true)
                             .execute()
             );
